+++ src/main/java/com/avg/lawsuitmanagement/lawsuit/repository/param/UpdateLawsuitInfoParam.java
@@ -15,8 +15,8 @@ public class UpdateLawsuitInfoParam {
     private String lawsuitType;
     private String name;
     private int courtId;
-    private int commissionFee;
-    private int contingentFee;
+    private Long commissionFee;
+    private Long contingentFee;
     private String lawsuitStatus;
     private String lawsuitNum;
     private String result;
